tidy stuff
