+++ source/cornell-box/cornellrenderer.ts
@@ -6,12 +6,8 @@ import {
 
 import { vec3, vec4 } from 'gl-matrix';
 
-// scene data
-import { indices, vertices } from './scene';
-
 // helper functions
-import { pointsOnSphere } from './icosphere';
-import { pointsInLight } from './light';
+import { pointsInLight, pointsOnSphere } from './helper';
 import { TrackballNavigation } from './trackballnavigation';
 
 
@@ -22,6 +18,10 @@ const _gCenter = vec3.fromValues(
     +0.000000, +0.000000, +0.000000);
 const _gUp = vec3.fromValues(
     +0.000000, +1.000000, +0.000000);
+
+// corners of axis aligned light cuboid
+const light0 = vec3.fromValues(-0.233813, +1 - 2e-4, -0.188126);
+const light1 = vec3.fromValues(+0.233813, +1 - 2e-4, +0.187411);
 
 
 export class CornellRenderer extends Renderer {
@@ -46,8 +46,6 @@ export class CornellRenderer extends Renderer {
     protected _uNdcOffset: WebGLUniformLocation;
 
     // Textures
-    // protected _verticesImage: Texture2;
-    protected _indicesImage: Texture2;
     protected _hsphereImage: Texture2;
     protected _lightsImage: Texture2;
 
@@ -136,10 +134,8 @@ export class CornellRenderer extends Renderer {
         gl.uniform1i(this._uRand, Math.floor(Math.random() * 1e6));
         gl.uniform2fv(this._uNdcOffset, ndcOffset);
 
-        // this._verticesImage.bind(gl.TEXTURE0);
-        this._indicesImage.bind(gl.TEXTURE0);
-        this._hsphereImage.bind(gl.TEXTURE1);
-        this._lightsImage.bind(gl.TEXTURE2);
+        this._hsphereImage.bind(gl.TEXTURE0);
+        this._lightsImage.bind(gl.TEXTURE1);
 
         // render geometry
         this._ndcTriangle.bind();
@@ -199,10 +195,8 @@ export class CornellRenderer extends Renderer {
         this._uViewport = this._program.uniform('u_viewport');
 
         this._program.bind();
-        // gl.uniform1i(this._program.uniform('u_vertices'), 0);
-        gl.uniform1i(this._program.uniform('u_indices'), 0);
-        gl.uniform1i(this._program.uniform('u_hsphere'), 1);
-        gl.uniform1i(this._program.uniform('u_lights'), 2);
+        gl.uniform1i(this._program.uniform('u_hsphere'), 0);
+        gl.uniform1i(this._program.uniform('u_lights'), 1);
         this._program.unbind();
 
         // triangle
@@ -210,28 +204,8 @@ export class CornellRenderer extends Renderer {
         const aVertex = this._program.attribute('a_vertex', 0);
         this._ndcTriangle.initialize(aVertex);
 
-        // scene textures
-        // this._verticesImage = new Texture2(this._context, 'verticesImage');
-        // this._verticesImage.initialize(vertices.length / 3, 1, gl.RGB32F, gl.RGB, gl.FLOAT); // height 1
-        // this._verticesImage.data(vertices);
-        // this._verticesImage.wrap(gl.CLAMP_TO_EDGE, gl.CLAMP_TO_EDGE);
-        // this._verticesImage.filter(gl.NEAREST, gl.NEAREST);
-
-        this._indicesImage = new Texture2(this._context, 'indicesImage');
-        this._indicesImage.initialize(indices.length / 4, 1, gl.RGBA8UI, gl.RGBA_INTEGER, gl.UNSIGNED_BYTE); // height 1
-        this._indicesImage.data(indices);
-        this._indicesImage.wrap(gl.CLAMP_TO_EDGE, gl.CLAMP_TO_EDGE);
-        this._indicesImage.filter(gl.NEAREST, gl.NEAREST);
-
-        // this._colorsImage = new Texture2(this._context, 'colorsImage');
-        // this._colorsImage.initialize(colors.length / 4
-        //     , 1, gl.RGBA32F, gl.RGBA, gl.FLOAT); // height 1
-        // this._colorsImage.data(colors);
-        // this._colorsImage.wrap(gl.CLAMP_TO_EDGE, gl.CLAMP_TO_EDGE);
-        // this._colorsImage.filter(gl.NEAREST, gl.NEAREST);
-
         // CREATE HEMISPHERE PATH SAMPLES
-        const points = pointsOnSphere(16384);
+        const points = pointsOnSphere(2048);
         const samplerSize = Math.floor(Math.sqrt(points.length));
         const spherePoints = new Float32Array(samplerSize * samplerSize * 3);
         for (let i = 0; i < samplerSize * samplerSize; ++i) {
@@ -246,9 +220,7 @@ export class CornellRenderer extends Renderer {
         this._hsphereImage.filter(gl.NEAREST, gl.NEAREST);
 
         // CREATE LIGHT AREA SAMPLES
-        const l1 = vec3.fromValues(vertices[0], vertices[1], vertices[2]);
-        const l2 = vec3.fromValues(vertices[6], vertices[7], vertices[8]);
-        const lights = pointsInLight(l1, l2, 32 * 32);
+        const lights = pointsInLight(light0, light1, 32 * 32);
         const lights2 = new Float32Array(lights.length * 3);
         let i2 = 0;
         for (const light of lights) {
@@ -292,9 +264,6 @@ export class CornellRenderer extends Renderer {
         this._program.uninitialize();
         this._ndcTriangle.uninitialize();
 
-        // this._verticesImage.uninitialize();
-        this._indicesImage.uninitialize();
-        // this._colorsImage.uninitialize();
         this._hsphereImage.uninitialize();
         this._lightsImage.uninitialize();
